Minor fixes and refactoring
- Formatting and refactoring

%SOFTWARE
