+++ numerical_methods_lib/src/lib.rs
@@ -1,6 +1,2 @@
-pub mod explicit_euler_method;
-pub mod heun_method;
-pub mod implicit_euler_method;
-pub mod ode_solver;
+pub mod ode_solvers;
 pub mod root_finders;
-pub mod runge_kutta4;
