+++ numerical_methods_lib/src/ode_solvers/explicit_euler_method.rs
@@ -1,4 +1,4 @@
-use crate::ode_solver::{OdeSolver, Printable, Solve};
+use crate::ode_solvers::ode_solver::{OdeSolver, Printable, Solve};
 
 /// Implements the Euler Method.
