+++ numerical_methods_lib/src/ode_solvers/mod.rs
@@ -1,0 +1,5 @@
+pub mod explicit_euler_method;
+pub mod heun_method;
+pub mod implicit_euler_method;
+pub mod ode_solver;
+pub mod runge_kutta4;
