+++ project_bin/src/main.rs
@@ -1,5 +1,5 @@
 use numerical_methods_lib::*;
-use ode_solver::{Printable, Solve};
+use ode_solvers::ode_solver::{Printable, Solve};
 
 const T_INITIAL: i32 = 0; // t0
 const T_FINAL: i32 = 1; // tf
@@ -15,16 +15,17 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     let f = |_x: f64, _y: f64| -20.0 * _x * _y * _y; // function: f(t,x)  // y marked as _y for now
     let f_dash = |_x: f64, _y: f64| -40.0 * _x * _y; // function: f'(t,x)  // y marked as _y for now
 
-    let solver_params = ode_solver::OdeSolverParams {
+    let solver_params = ode_solvers::ode_solver::OdeSolverParams {
         f: f,
         num_steps: num_steps,
         t_initial: T_INITIAL,
         time_step: TIME_STEP,
         ..Default::default() // We do not need the rest for the explicit method, so we leave as Default
     };
-    let solver_object = ode_solver::OdeSolver::new("ODE Solver Explicit Euler", &solver_params);
+    let solver_object =
+        ode_solvers::ode_solver::OdeSolver::new("ODE Solver Explicit Euler", &solver_params);
 
-    let explicit_euler_solver = explicit_euler_method::ExplicitEulerSolver {
+    let explicit_euler_solver = ode_solvers::explicit_euler_method::ExplicitEulerSolver {
         solver: Box::new(solver_object),
     };
     explicit_euler_solver.solve(&mut solution);
@@ -32,9 +33,9 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     solution.clear();
     solution.push(INITIAL_SOLUTION);
 
-    let solver_object2 = ode_solver::OdeSolver::new("ODE Solver Heun", &solver_params);
+    let solver_object2 = ode_solvers::ode_solver::OdeSolver::new("ODE Solver Heun", &solver_params);
 
-    let heun_method_solver = heun_method::HeunSolver {
+    let heun_method_solver = ode_solvers::heun_method::HeunSolver {
         solver: Box::new(solver_object2),
     };
     heun_method_solver.solve(&mut solution);
@@ -42,9 +43,10 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     solution.clear();
     solution.push(INITIAL_SOLUTION);
 
-    let solver3 = ode_solver::OdeSolver::new("ODE Solver Runge Kutta 4", &solver_params);
+    let solver3 =
+        ode_solvers::ode_solver::OdeSolver::new("ODE Solver Runge Kutta 4", &solver_params);
 
-    let rungekutta_solver = runge_kutta4::RungeKuttaSolver {
+    let rungekutta_solver = ode_solvers::runge_kutta4::RungeKuttaSolver {
         solver: Box::new(solver3),
     };
     rungekutta_solver.solve(&mut solution);
@@ -52,7 +54,7 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
     solution.clear();
     solution.push(INITIAL_SOLUTION);
 
-    let solver_params_implicit = ode_solver::OdeSolverParams {
+    let solver_params_implicit = ode_solvers::ode_solver::OdeSolverParams {
         f,
         f_dash,
         num_steps,
@@ -60,8 +62,11 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
         max_iters: MAX_ITERATIONS,
         ..Default::default() // We can fill all the values, but for this case I have just filled the ones extra for the implicit
     };
-    let solver4 = ode_solver::OdeSolver::new("ODE Solver Implicit Euler", &solver_params_implicit);
-    let implicit_euler_solver = implicit_euler_method::ImplicitEulerSolver {
+    let solver4 = ode_solvers::ode_solver::OdeSolver::new(
+        "ODE Solver Implicit Euler",
+        &solver_params_implicit,
+    );
+    let implicit_euler_solver = ode_solvers::implicit_euler_method::ImplicitEulerSolver {
         solver: Box::new(solver4),
     };
     implicit_euler_solver.solve(&mut solution);
